Worked on issue #107, test app for edge cases: if a user is on an unsaved day and then goes to another unsaved day. Does note data persist on the correct days?

+++ fitnessjournal/src/pages/WorkoutLog.jsx
@@ -346,7 +346,7 @@ export default function WorkoutLog() {
         e.preventDefault()
         const workoutData = !alreadySavedWorkout ? JSON.parse(localStorage.getItem('exercises')):
         JSON.parse(localStorage.getItem('workoutData'))
-        console.log(alreadySavedWorkout)
+        
         if(!alreadySavedWorkout) {
             const updatedWorkoutData = workoutData.map(exercise => {
                 // check to see if exercise matches. If it does, continue with deletion
@@ -379,28 +379,45 @@ export default function WorkoutLog() {
     
     function addSetNote(e) {
         e.preventDefault()
-        const workoutData = JSON.parse(localStorage.getItem('exercises'))
-        const updatedWorkoutData = workoutData.map(exercise => {
-            // check to see if exercise matches. If it does, continue with edit
-            if(exercise.id === newSetInfo.exId) {
-                // creates a shallow copy of setsReps array in given exercise
-                const updatedSetsReps = [...exercise.setsReps]
-                // updates set at given setIndex with newSetInfo
-                updatedSetsReps[newSetInfo.setIndex] = {
-                    ...updatedSetsReps[newSetInfo.setIndex],
-                    note: note
+        const workoutData = !alreadySavedWorkout ? JSON.parse(localStorage.getItem('exercises')):
+        JSON.parse(localStorage.getItem('workoutData'))
+
+        if(!alreadySavedWorkout) {
+            const updatedWorkoutData = workoutData.map(exercise => {
+                // check to see if exercise matches. If it does, continue with edit
+                if(exercise.id === newSetInfo.exId) {
+                    // creates a shallow copy of setsReps array in given exercise
+                    const updatedSetsReps = [...exercise.setsReps]
+                    // updates set at given setIndex with newSetInfo
+                    updatedSetsReps[newSetInfo.setIndex] = {
+                        ...updatedSetsReps[newSetInfo.setIndex],
+                        note: note
+                    }
+                    // returns the exercise info plus the updated sets
+                    return {
+                        ...exercise,
+                        setsReps: updatedSetsReps
+                    }
                 }
-                // returns the exercise info plus the updated sets
-                return {
-                    ...exercise,
-                    setsReps: updatedSetsReps
-                }
+                
+                // if exercise id doesn't match the exid of the set the user clicks,
+                // returns the unchanged exercise so that other exercises remain the same.
+                return exercise
+            })
+            localStorage.setItem('exercises', JSON.stringify(updatedWorkoutData))
+        } else if(alreadySavedWorkout) {
+            try {
+                workoutData.forEach(exercise => {
+                    if(exercise.id === newSetInfo.exId) {
+                        console.log(exercise.id, newSetInfo.exId, newSetInfo.setId)
+                        AddSetNote(currentUser, usersInDB, date, exercise.id, newSetInfo.setId, note)
+                    }
+                })
+                
+            } catch(err) {
+                console.error('error saving note: ', err)
             }
-            // if exercise id doesn't match the exid of the set the user clicks,
-            // returns the unchanged exercise so that other exercises remain the same.
-            return exercise
-        })
-        localStorage.setItem('exercises', JSON.stringify(updatedWorkoutData))
+        }
         
         loadExerciseList()
         setCurrentNote(note)
@@ -552,7 +569,7 @@ export default function WorkoutLog() {
                     modalStyles={modalStyles}
                 />
             }
-
+   
             {
                 toggleNoteForm &&
                 <AddNote
